Re-establish original behaviour of Read() method
The original Read() method was a wrapper to C.hid_read.
C.hid_read checks for the device blocking field
to pass a timeout value to C.hid_read_timeout accordingly:
if blocking is enabled, -1, else 0.

The changed introduced in https://github.com/karalabe/hid/pull/22
and merged in this fork breaks that and incorrectly assumes that
a 0 timeout means a blocking read. This commit allows to maintain
the original use of Read() while also allowing a package user to
make use of ReadTimeout(b, t)
ReadTimeout() was previously missing from hid_disable.go